emacs: Reimplement image-dired database with hash-table
This commit reimplements image-dired tag (and comment) database with a more
reliable one.  Before, it used a simple plain-text database in the following
form:

/file/path;tag1;tag2;comment:my comment message
/file/path;tag1;tag2;comment:my comment message

This is very error prone.  A comment can be very easily confused for a tag.  A
file name with characters that needs escaping were not handled properly.

Now, the database is a serialized hash-table.  It is very easy to search and
modify its content.  However, maybe, problems with performance could arise.  I
have not properly tested this implementations with large datasets.